+++ src/Api/Dfe.Complete.Api/Controllers/ProjectsController.cs
@@ -62,6 +62,37 @@ namespace Dfe.Complete.Api.Controllers
         }
 
         /// <summary>
+        /// Creates a new MAT conversion project (handover version).
+        /// </summary>
+        /// <param name="request">The request.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWrite")]
+        [HttpPost]
+        [Route("projects/conversions/form-a-mat")]
+        [SwaggerResponse(201, "Project created successfully.", typeof(ProjectId))]
+        [SwaggerResponse(400, "Invalid request data.")]
+        public async Task<IActionResult> CreateHandoverConversionMatProjectAsync([FromBody] CreateHandoverConversionMatProjectCommand request, CancellationToken cancellationToken)
+        {
+            try
+            {
+                var result = await sender.Send(request, cancellationToken);
+                return Created("", result);
+            }
+            catch (UnprocessableContentException ex)
+            {
+                return UnprocessableEntity(ex.Message);
+            }
+            catch (ValidationException ex)
+            {
+                return BadRequest(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, ex.Message);
+            }
+        }
+
+        /// <summary>
         /// Creates a new transfer project (handover version).
         /// </summary>
         /// <param name="request">The request.</param>