+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Security in case of vulnerabilities.
 - Added `Share the information about the opening' task for conversion project.
 - Added `Subleases` task
 - Added `Tenancy at will` task for conversion project.
+- Added project group creation end point
+- Added MAT Conversion creation end point for prepare to complete 
 
 ### Changed
 
@@ -66,8 +68,6 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - Added 'Land consent letter task for transfer projects'
 - Adding key contact record on confirming project handover and logging error message if key contact is already been added.
 - Added `Check accuracy of high needs places information` task for conversion project.
-- Added 'Land consent letter task for transfer projects'
-- Added project group creation end point
 
 ### Fixed